✨ feat : 여행방 현금입력 구현 & 방 정보 세팅 구현 | 하동혁

+++ Travelance/data/src/main/java/com/moneyminions/data/model/home/response/TravelRoomInfoResponse.kt
@@ -11,7 +11,7 @@ data class TravelRoomInfoResponse(
     @SerializedName("budget")
     var budget: Int,
     @SerializedName("percent")
-    var percent: Int,
+    var percent: Float,
     @SerializedName("rest")
     var rest: Int,
     @SerializedName("isDone")
